refactor. main & storeList
